Server/requirements.txt
bcrypt
flask
flasgger
flask-cors
passlib
pymongo
pydantic
PyJWT
